now using androidx + gradle, using device mac address as identifier
