--- android/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Mar 04 17:05:22 CET 2019
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.10.1-all.zip
