+++ app/src/main/res/layout/activity_main.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".MainActivity">
+
+    <TextView
+            android:layout_width="210dp"
+            android:layout_height="20dp"
+            android:id="@+id/textView"
+            android:text="@string/printer_select"
+            app:layout_constraintTop_toTopOf="parent"
+            android:layout_marginTop="36dp" app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"/>
+    <Button
+            android:text="@string/test_print"
+            android:layout_width="210dp"
+            android:layout_height="50dp"
+            android:id="@+id/testPrintButton"
+            android:layout_marginTop="8dp"
+            app:layout_constraintTop_toBottomOf="@+id/printerSpinner" app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:background="@color/colorPrimary" android:textColor="@color/colorWhite"/>
+    <Spinner
+            style="@style/Widget.AppCompat.Spinner.Underlined"
+            android:layout_width="210dp"
+            android:layout_height="50dp"
+            android:id="@+id/printerSpinner"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginTop="8dp" app:layout_constraintTop_toBottomOf="@+id/textView"
+            android:spinnerMode="dropdown"/>
+    <Button
+            android:text="@string/bt_scan"
+            android:layout_width="210dp"
+            android:layout_height="50dp"
+            android:id="@+id/scanButton" android:layout_marginTop="8dp"
+            app:layout_constraintTop_toBottomOf="@+id/testPrintButton" app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:textColor="@color/colorDarkText"/>
+    <TextView
+            android:text="2020 © Ethereals United"
+            android:layout_width="wrap_content"
+            android:layout_height="36dp"
+            android:id="@+id/textView3" android:layout_marginTop="8dp"
+            app:layout_constraintTop_toBottomOf="@+id/scanButton"
+            app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintEnd_toEndOf="parent"
+            android:layout_marginEnd="8dp" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="8dp" android:gravity="center" android:layout_marginBottom="8dp"
+            app:layout_constraintHorizontal_bias="0.497" app:layout_constraintVertical_bias="0.846"/>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
