--- www/index.html
@@ -1,16 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-    <title>GOKASA Development</title>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="refresh" content="0; url=/gokasa-printer.html" />
-</head>
-
-<body>
-    <p><a href="/gokasa-printer.html">Redirect</a></p>
-</body>
-
-</html>
